+++ internal/schemas/balance.go
@@ -8,8 +8,8 @@ type Balance struct {
 }
 
 type Withdraw struct {
-	Order string `json:"order"`
-	Sum   int    `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float32 `json:"sum"`
 }
 
 type Withdrawal struct {
